+++ build.php
@@ -2,7 +2,7 @@
 /**
  * extension source version
  */
-$version = '1.2.0';
+$version = '1.3.0';
 
 /**
  * dist filename